requirements.txt
Django>=1.9
djangorestframework